+++ Tests/AdvancedOperationTests/AdvancedBlockOperationTests.swift
@@ -32,7 +32,9 @@ final class AdvancedBlockOperationTests: XCTestCase {
         complete([])
       }
     }
-
+    XCTAssertTrue(operation.isAsynchronous)
+    XCTAssertTrue(operation.isConcurrent)
+    
     let expectation1 = XCTKVOExpectation(keyPath: #keyPath(AdvancedOperation.isFinished), object: operation, expectedValue: true)
     operation.start()
     operation.cancel()