+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Changelog
 
-### 2.0.0 (WIP)
+### 2.0.0
 
 - Added `BlockCondition`. 
 - Added `TimeoutObserver`.
+- Added support for `isAsynchronous`  \ `isCondurrent` operation.
 - Fixed KVO for the `iSCancelled` property in `AdvancedOperation`.
 - A `GroupOperation` can be cancelled correctly even if it's not yet started.
 - The propery `failed` in `AdvancedOperation` has been renamed `hasErrors` to better reflect its intent.